requirements.txt
flask==3.0.3
streamlit==1.40.0
transformers==4.46.0
torch>=2.2.0
sentence-transformers==3.2.0
faiss-cpu==1.13.2
python-dotenv==1.0.1
numpy>=2.1.0
scikit-learn==1.5.2
requests==2.32.3
